Switch to SDSNA
Switch from the `NumberArray` derivative type to the new derivative
type `SemiDynamicSparseNumberArray`. This template class has
data members to hold both the derivative values and the derivative
indices, meaning that we can take greatest advantage of sparsity.
The backing storage type is `std::array` so that storage is static
and we don't see any slow down from dynamic memory allocation.
We avoid looping over the whole size of the std::array by
maintaining a dynamic size data member. The only real draw-back
to this derivative type is that we'll hit an assertion if our sparse
size ever exceeds the static storage size of the std::array. You can get
around this by increasing the static size of the array, but you will
incur some (small?) memory overhead because of the material properties
and variable values (dynamic storage containers std::vector and MooseArray
respectively) that hold MOOSE's DualReal objects.